School-Management-System/requirements.txt
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
connect==0.2
dnspython==2.3.0
Flask==2.3.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
protobuf==3.20.3
psycopg2==2.9.6
pymango==0.1.1
pymongo==4.4.0
python-geoip==1.2
python-geoip-geolite2==2015.303
requests==2.31.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
urwid==2.1.2
Werkzeug==2.3.4
wtf==0.1
WTForms==3.0.1